+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 GAMSPy CHANGELOG
 ================
 
+GAMSPy 1.5.0
+------------
+- General
+  - Fix implicit parameter validation bug.
+  - Allow the usage of Container as a context manager.
+  - Allow propagating bounds to the output variable in `flatten_dims` method.
+  - Add piecewise linear function formulations.
+  - Migrate GAMSPy CLI to Typer.
+  - Threads can now create a container since we register the signal only to the main thread.
+  - Fix solver options bug in frozen solve.
+  - Synchronize after read.
+  - Upgrade gamspy_base and gamsapi dependencies.
+  - Add `--checkout-duration` and `--renew` options to `gamspy install license`.
+- Testing
+  - Lower the number of dices in the interrupt test and put a time limit to the solve.
+  - Add tests for piecewise linear functions.
+- Documentation
+  - Install dependencies in the first cell of the example transportation notebook.
+  - Add Formulations page to list piecewise linear functions and nn formulations.
+
 GAMSPy 1.4.0
 ------------
 - General
@@ -17,6 +37,7 @@ GAMSPy 1.4.0
   - Remove non-negative variable type from the docs.
   - Add plausible.js for analytics.
   - Minor update in embedding nn documentation.
+  - Add descriptions and example code to formulations documentation.
 
 
 GAMSPy 1.3.1